+++ functional/slices/union.go
@@ -2,17 +2,19 @@ package slices
 
 import "github.com/couchbase/tools-common/functional/maps"
 
-// Union returns a slice of elements that are present in both input slices.
+// Union returns all the elements that are in either 'a' or 'b'.
 //
-// NOTE: The returned slice will not contain any duplicates.
+// NOTE: The provided slices may contain duplicates but the returned slice will not.
 func Union[S []E, E comparable](a, b S) S {
-	union := make(map[E]uint64)
+	union := make(map[E]struct{})
 
-	for _, slice := range []S{a, b} {
-		for _, v := range slice {
-			union[v]++
-		}
+	for _, e := range a {
+		union[e] = struct{}{}
 	}
 
-	return maps.Keys(union, func(_ E, v uint64) bool { return v > 1 })
+	for _, e := range b {
+		union[e] = struct{}{}
+	}
+
+	return maps.Keys(union)
 }